test-requirements.txt
pytz>=2013.6  # MIT
PyYAML>=3.1.0  # MIT
yaml2ical>=0.6.1
